+++ python/ray/train/__init__.py
@@ -1,9 +1,9 @@
 from ray._private.usage import usage_lib
-from ray.train.backend import BackendConfig
-from ray.train.data_config import DataConfig
-from ray.train.context import get_context
-from ray.train.constants import TRAIN_DATASET_KEY
+from ray.train._internal.data_config import DataConfig
 from ray.train._internal.session import get_dataset_shard, report
+from ray.train.backend import BackendConfig
+from ray.train.constants import TRAIN_DATASET_KEY
+from ray.train.context import get_context
 from ray.train.trainer import TrainingIterator
 
 from ray.air import Checkpoint
